agent/requirements.txt
# LangGraph and LangChain
langchain>=1.0.0
langchain-core>=1.0.0
langgraph>=1.0.0
langchain-deepseek>=1.0.0

# HTTP Client
aiohttp>=3.9.0

# MCP Protocol (for backward compatibility)
mcp>=1.0.0

# Async support
asyncio-throttle>=1.0.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
